+++ src/main/java/me/universi/user/services/CustomUsernamePasswordAuthenticationFilter.java
@@ -67,7 +67,7 @@ public class CustomUsernamePasswordAuthenticationFilter extends UsernamePassword
                 this.jsonUsername = (String)( mapRequest.containsKey("username")? mapRequest.get("username") : mapRequest.get("email"));
                 this.jsonPassword = (String)mapRequest.get("password");
 
-                UserService.getInstance().checkRecaptchaWithToken(mapRequest.get("recaptchaToken"));
+                GoogleService.getInstance().checkRecaptchaWithToken(mapRequest.get("recaptchaToken"));
 
             } catch (Exception e) {
                 try {
